Added test for DotNetCommandProjectPathUpdaterTests

+++ src/MeerkatUpdater/MeerkatUpdater.Core.Test/GeneralUsage/CreatedFoldersManager.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Globalization;
 using System.IO;
+using System.Threading;
 
 namespace MeerkatUpdater.Core.Test.GeneralUsage
 {
@@ -20,10 +21,26 @@ namespace MeerkatUpdater.Core.Test.GeneralUsage
             return folder;
         }
 
-        public static void TierDownTest(string testKey)
+        public static void TierDownTest(string testKey, int retryCount = 0)
         {
+            const int MaximumRetryCount = 3;
+
+            if (retryCount > MaximumRetryCount)
+                return;
+
             if (ListOfCreatedFolders.TryGetValue(testKey, out string path))
-                Directory.Delete(path, true);
+            {
+                try
+                {
+                    Directory.Delete(path, true);
+                }
+                catch
+                {
+                    Thread.Sleep(TimeSpan.FromSeconds(3));
+                    retryCount++;
+                    TierDownTest(testKey, retryCount);
+                }
+            }
         }
 
         private static int GetRandomNumber(int min, int max)
